필요 없는 형변환 제거 -> (int) 제거

+++ BAEKJOON/2_week/yejin/Q2164.java
@@ -16,9 +16,9 @@ public class Q2164 {
         }
         while(que.size()!=1){
             que.poll();
-            que.add((int)que.poll());
+            que.add(que.poll());
         }
-        return (int)que.poll();
+        return que.poll();
     }
 
     public static void main(String[] args) {
